[via Eronarn] monsters.py for generating data lists about all monsters
Obviously a rather large part of Crawl's monsters still is not handled
properly or at all with this, for instance derived undead or ghost_demon
data. Additionally, many resistances that are set dynamically in Crawl
cannot be parsed from mon-data.h exclusively at this point.

I tried to replicate some of that functionality by looking at holiness
for the less complicated cases, but that solution is far from complete.